@@ -3,7 +3,7 @@
   "homepage": "https://github.com/k-holy/volcanus-query-builder",
   "type": "library",
   "description": "SQL builder for volcanus-database",
-  "version": "2.1.1",
+  "version": "3.0.0",
   "license": "MIT",
   "authors": [
     {
@@ -17,8 +17,8 @@
     }
   },
   "require": {
-    "php": ">=7.3||>=8.0.2",
-    "volcanus/database": "^2.1"
+    "php": "^8.1",
+    "volcanus/database": "^3.0"
   },
   "require-dev": {
     "phpunit/phpunit": "^9.5",
