@@ -1,6 +1,6 @@
 <?php
 /**
- * Volcanus libraries for PHP
+ * Volcanus libraries for PHP 8.1~
  *
  * @copyright k-holy <k.holy74@gmail.com>
  * @license The MIT License (MIT)
@@ -20,19 +20,9 @@ class MysqlQueryBuilder extends AbstractQueryBuilder implements QueryBuilderInte
 {
 
     /**
-     * @var MysqlExpressionBuilder
-     */
-    protected $expressionBuilder;
-
-    /**
-     * @var MysqlParameterBuilder
-     */
-    protected $parameterBuilder;
-
-    /**
      * @var array サポートするデータ型名
      */
-    protected static $types = [
+    protected static array $types = [
         'text' => ['text', 'char', 'varchar', 'tinytext', 'longtext', 'mediumtext', 'json'],
         'int' => ['int', 'integer', 'tinyint', 'int4', 'smallint', 'mediumint', 'bigint'],
         'float' => ['decimal', 'float', 'double', 'real'],
@@ -63,7 +53,7 @@ class MysqlQueryBuilder extends AbstractQueryBuilder implements QueryBuilderInte
      * @param int|null $offset 取得開始行index
      * @return string SQL
      */
-    public function limitOffset(string $sql, ?int $limit = null, ?int $offset = null): string
+    public function limitOffset(string $sql, int $limit = null, int $offset = null): string
     {
         return sprintf('%s LIMIT %s%s',
             $sql,
@@ -80,7 +70,7 @@ class MysqlQueryBuilder extends AbstractQueryBuilder implements QueryBuilderInte
      */
     public function count(string $sql): string
     {
-        if (false !== strpos($sql, 'SQL_CALC_FOUND_ROWS')) {
+        if (str_contains($sql, 'SQL_CALC_FOUND_ROWS')) {
             return 'SELECT FOUND_ROWS()';
         }
         return sprintf('SELECT COUNT(*) FROM (%s) AS __SUBQUERY', $sql);
