@@ -1,6 +1,6 @@
 <?php
 /**
- * Volcanus libraries for PHP
+ * Volcanus libraries for PHP 8.1~
  *
  * @copyright k-holy <k.holy74@gmail.com>
  * @license The MIT License (MIT)
@@ -20,17 +20,17 @@ class Facade
     /**
      * @var DriverInterface
      */
-    private $driver;
+    private DriverInterface $driver;
 
     /**
      * @var QueryBuilderInterface
      */
-    private $builder;
+    private QueryBuilderInterface $builder;
 
     /**
-     * @var boolean 取得するカラム名をキャメルケースに変換するかどうか
+     * @var bool 取得するカラム名をキャメルケースに変換するかどうか
      */
-    private $enableCamelize = false;
+    private bool $enableCamelize = false;
 
     /**
      * Facade constructor.
@@ -48,8 +48,9 @@ class Facade
      * 取得するカラム名をキャメルケースに変換するかどうかを設定します。
      *
      * @param mixed $enabled 取得するカラム名をキャメルケースに変換するかどうか
+     * @return void
      */
-    public function enableCamelize($enabled = true)
+    public function enableCamelize(mixed $enabled = true): void
     {
         $this->enableCamelize = (bool)$enabled;
     }
@@ -62,7 +63,7 @@ class Facade
      * @param string|null $alias 別名
      * @return string SQL句
      */
-    public function expression(string $expr, ?string $type = null, ?string $alias = null): string
+    public function expression(string $expr, string $type = null, string $alias = null): string
     {
         return $this->builder->expression($expr, $type, $alias);
     }
@@ -76,7 +77,7 @@ class Facade
      * @param array|null $columnAliases 別名取得設定 (キー=列名、値=別名)
      * @return array SQL SELECT句となる列指定の配列
      */
-    public function expressions(string $tableName, ?string $tableAlias = null, ?array $excludeKeys = [], ?array $columnAliases = []): array
+    public function expressions(string $tableName, string $tableAlias = null, ?array $excludeKeys = [], ?array $columnAliases = []): array
     {
         $metaColumns = $this->driver->getMetaColumns($tableName);
         $expressions = [];
@@ -101,7 +102,7 @@ class Facade
      * @param string $type 型名 ($typesフィールド参照)
      * @return string 変換結果
      */
-    public function parameter($value, string $type): string
+    public function parameter(mixed $value, string $type): string
     {
         return $this->builder->parameter($value, $type);
     }
@@ -239,7 +240,7 @@ SQL;
      * @param array|null $columnAliases 別名取得設定 (キー=列名、値=別名)
      * @return string SQL
      */
-    public function select(string $tableName, ?string $tableAlias = null, ?string $where = null, ?array $excludeKeys = [], ?array $columnAliases = []): string
+    public function select(string $tableName, string $tableAlias = null, string $where = null, ?array $excludeKeys = [], ?array $columnAliases = []): string
     {
         $sql = join("\n", [
             $this->selectSyntax($tableName, $tableAlias, $excludeKeys, $columnAliases),
@@ -269,7 +270,7 @@ SQL;
      * @param int|null $offset 取得開始行index
      * @return string SQL
      */
-    public function limitOffset(string $sql, ?int $limit = null, ?int $offset = null): string
+    public function limitOffset(string $sql, int $limit = null, int $offset = null): string
     {
         return $this->builder->limitOffset($sql, $limit, $offset);
     }
@@ -282,7 +283,7 @@ SQL;
      * @param array|null $columns 抽出条件を格納した配列(キー=列名、値=列値)
      * @return array SQL WHERE句となる抽出条件の配列
      */
-    public function whereExpressions(string $tableName, ?string $tableAlias = null, ?array $columns = []): array
+    public function whereExpressions(string $tableName, string $tableAlias = null, ?array $columns = []): array
     {
         if (empty($columns)) {
             return [];
@@ -359,7 +360,7 @@ SQL;
      * @param array|null $orders 整列順を格納した配列(キー=順序、値=列名 + 昇順/降順)
      * @return array SQL ORDER BY句となる列名の配列
      */
-    public function orderByExpressions(string $tableName = null, ?string $tableAlias = null, ?array $orders = []): array
+    public function orderByExpressions(string $tableName = null, string $tableAlias = null, ?array $orders = []): array
     {
         if (empty($orders)) {
             return [];
@@ -398,7 +399,7 @@ SQL;
      * @param array|null $appendKeys 追加列を格納した配列(値=列名)
      * @return array SQL GROUP BY句となる列名の配列
      */
-    public function groupByExpressions(string $tableName, ?string $tableAlias = null, ?array $excludeKeys = [], ?array $appendKeys = []): array
+    public function groupByExpressions(string $tableName, string $tableAlias = null, ?array $excludeKeys = [], ?array $appendKeys = []): array
     {
         $metaColumns = $this->driver->getMetaColumns($tableName);
         $expressions = [];
@@ -423,7 +424,7 @@ SQL;
      * @param array|null $columns 抽出条件を格納した配列(キー=列名、値=列値)
      * @return string
      */
-    public function whereSyntax(string $tableName, ?string $tableAlias = null, ?array $columns = []): string
+    public function whereSyntax(string $tableName, string $tableAlias = null, ?array $columns = []): string
     {
         $expressions = $this->whereExpressions($tableName, $tableAlias, $columns);
         $where = (count($expressions) > 0) ? join(" AND\n", $expressions) : '';
@@ -438,7 +439,7 @@ SQL;
      * @param array|null $orders 整列順を格納した配列(キー=順序、値=列名 + 昇順/降順)
      * @return string
      */
-    public function orderBySyntax(string $tableName, ?string $tableAlias = null, ?array $orders = []): string
+    public function orderBySyntax(string $tableName, string $tableAlias = null, ?array $orders = []): string
     {
         $expressions = $this->orderByExpressions($tableName, $tableAlias, $orders);
         $orderBy = (count($expressions) > 0) ? join(" ,\n", $expressions) : '';
@@ -454,7 +455,7 @@ SQL;
      * @param array|null $appendKeys 追加列を格納した配列(値=列名)
      * @return string
      */
-    public function groupBySyntax(string $tableName, ?string $tableAlias = null, ?array $excludeKeys = [], ?array $appendKeys = []): string
+    public function groupBySyntax(string $tableName, string $tableAlias = null, ?array $excludeKeys = [], ?array $appendKeys = []): string
     {
         $expressions = $this->groupByExpressions($tableName, $tableAlias, $excludeKeys, $appendKeys);
         $groupBy = (count($expressions) > 0) ? join(" ,\n", $expressions) : '';
