@@ -1,6 +1,6 @@
 <?php
 /**
- * Volcanus libraries for PHP
+ * Volcanus libraries for PHP 8.1~
  *
  * @copyright k-holy <k.holy74@gmail.com>
  * @license The MIT License (MIT)
@@ -20,19 +20,9 @@ class SqliteQueryBuilder extends AbstractQueryBuilder implements QueryBuilderInt
 {
 
     /**
-     * @var SqliteExpressionBuilder
-     */
-    protected $expressionBuilder;
-
-    /**
-     * @var SqliteParameterBuilder
-     */
-    protected $parameterBuilder;
-
-    /**
      * @var array サポートするデータ型名
      */
-    protected static $types = [
+    protected static array $types = [
         'text' => [
             'character',
             'varchar',
@@ -70,7 +60,7 @@ class SqliteQueryBuilder extends AbstractQueryBuilder implements QueryBuilderInt
      * @param int|null $offset 取得開始行index
      * @return string SQL
      */
-    public function limitOffset(string $sql, ?int $limit = null, ?int $offset = null): string
+    public function limitOffset(string $sql, int $limit = null, int $offset = null): string
     {
         $sql .= sprintf(' LIMIT %s',
             (isset($limit) && $limit >= 0)
