@@ -1,6 +1,6 @@
 <?php
 /**
- * Volcanus libraries for PHP
+ * Volcanus libraries for PHP 8.1~
  *
  * @copyright k-holy <k.holy74@gmail.com>
  * @license The MIT License (MIT)
@@ -22,7 +22,7 @@ interface ParameterBuilderInterface
      * @param mixed $value 値
      * @return string 変換結果
      */
-    public function toText($value): string;
+    public function toText(mixed $value): string;
 
     /**
      * 値を数値を表すSQLパラメータ値に変換します。
@@ -31,7 +31,7 @@ interface ParameterBuilderInterface
      * @param string|null $type 型名 ($typesフィールド参照)
      * @return string 変換結果
      */
-    public function toInt($value, ?string $type = null): string;
+    public function toInt(mixed $value, string $type = null): string;
 
     /**
      * 値を浮動小数点数を表すSQLパラメータ値に変換します。
@@ -40,7 +40,7 @@ interface ParameterBuilderInterface
      * @param string|null $type 型名 ($typesフィールド参照)
      * @return string 変換結果
      */
-    public function toFloat($value, ?string $type = null): string;
+    public function toFloat(mixed $value, string $type = null): string;
 
     /**
      * 値を真偽値を表すSQLパラメータ値に変換します。
@@ -48,7 +48,7 @@ interface ParameterBuilderInterface
      * @param mixed $value 値
      * @return string 変換結果
      */
-    public function toBool($value): string;
+    public function toBool(mixed $value): string;
 
     /**
      * 値を日付を表すSQLパラメータ値に変換します。
@@ -56,7 +56,7 @@ interface ParameterBuilderInterface
      * @param mixed $value 値
      * @return string 変換結果
      */
-    public function toDate($value): string;
+    public function toDate(mixed $value): string;
 
     /**
      * 値を日時を表すSQLパラメータ値に変換します。
@@ -64,6 +64,6 @@ interface ParameterBuilderInterface
      * @param mixed $value 値
      * @return string 変換結果
      */
-    public function toTimestamp($value): string;
+    public function toTimestamp(mixed $value): string;
 
 }
