@@ -1,6 +1,6 @@
 <?php
 /**
- * Volcanus libraries for PHP
+ * Volcanus libraries for PHP 8.1~
  *
  * @copyright k-holy <k.holy74@gmail.com>
  * @license The MIT License (MIT)
@@ -23,9 +23,9 @@ use Volcanus\Database\MetaData\MysqlMetaDataProcessor;
 class MysqlParameterBuilderTest extends TestCase
 {
 
-    private static $pdo;
+    private static \PDO $pdo;
 
-    public function getPdo()
+    private function getPdo(): \PDO
     {
         if (!isset(static::$pdo)) {
             static::$pdo = new \PDO('sqlite::memory:');
@@ -417,7 +417,6 @@ class MysqlParameterBuilderTest extends TestCase
         $builder = new MysqlParameterBuilder(
             new PdoDriver($this->getPdo(), new MysqlMetaDataProcessor())
         );
-        /** @noinspection PhpParamsInspection */
         $builder->ToDate(new \stdClass());
     }
 
@@ -545,7 +544,6 @@ class MysqlParameterBuilderTest extends TestCase
         $builder = new MysqlParameterBuilder(
             new PdoDriver($this->getPdo(), new MysqlMetaDataProcessor())
         );
-        /** @noinspection PhpParamsInspection */
         $builder->toTimestamp(new \stdClass());
     }
 
@@ -652,7 +650,6 @@ class MysqlParameterBuilderTest extends TestCase
         $builder = new MysqlParameterBuilder(
             new PdoDriver($this->getPdo(), new MysqlMetaDataProcessor())
         );
-        /** @noinspection PhpParamsInspection */
         $builder->toTime(new \stdClass());
     }
 
@@ -759,7 +756,6 @@ class MysqlParameterBuilderTest extends TestCase
         $builder = new MysqlParameterBuilder(
             new PdoDriver($this->getPdo(), new MysqlMetaDataProcessor())
         );
-        /** @noinspection PhpParamsInspection */
         $builder->toGeometry(new \stdClass());
     }
 
