@@ -1,6 +1,6 @@
 <?php
 /**
- * Volcanus libraries for PHP
+ * Volcanus libraries for PHP 8.1~
  *
  * @copyright k-holy <k.holy74@gmail.com>
  * @license The MIT License (MIT)
@@ -23,9 +23,9 @@ use Volcanus\Database\MetaData\SqliteMetaDataProcessor;
 class SqliteParameterBuilderTest extends TestCase
 {
 
-    private static $pdo;
+    private static \PDO $pdo;
 
-    public function getPdo()
+    private function getPdo(): \PDO
     {
         if (!isset(static::$pdo)) {
             static::$pdo = new \PDO('sqlite::memory:');
@@ -339,7 +339,6 @@ class SqliteParameterBuilderTest extends TestCase
         $builder = new SqliteParameterBuilder(
             new PdoDriver($this->getPdo(), new SqliteMetaDataProcessor())
         );
-        /** @noinspection PhpParamsInspection */
         $builder->ToDate(new \stdClass());
     }
 
@@ -452,7 +451,6 @@ class SqliteParameterBuilderTest extends TestCase
         $builder = new SqliteParameterBuilder(
             new PdoDriver($this->getPdo(), new SqliteMetaDataProcessor())
         );
-        /** @noinspection PhpParamsInspection */
         $builder->toTimestamp(new \stdClass());
     }
 
