@@ -1,6 +1,6 @@
 <?php
 /**
- * Volcanus libraries for PHP
+ * Volcanus libraries for PHP 8.1~
  *
  * @copyright k-holy <k.holy74@gmail.com>
  * @license The MIT License (MIT)
@@ -28,7 +28,7 @@ class FacadeTest extends TestCase
 {
 
     /** @var \PDO */
-    private static $pdo;
+    private static \PDO $pdo;
 
     private function getPdo(): \PDO
     {
