@@ -1,6 +1,6 @@
 <?php
 /**
- * Volcanus libraries for PHP
+ * Volcanus libraries for PHP 8.1~
  *
  * @copyright k-holy <k.holy74@gmail.com>
  * @license The MIT License (MIT)
@@ -11,6 +11,8 @@ namespace Volcanus\QueryBuilder\Test;
 use PHPUnit\Framework\TestCase;
 use Volcanus\QueryBuilder\QueryBuilder;
 use Volcanus\Database\Driver\Pdo\PdoDriver;
+use Volcanus\QueryBuilder\Facade;
+use Volcanus\QueryBuilder\Adapter\Sqlite\SqliteQueryBuilder;
 
 /**
  * Test for QueryBuilder
@@ -22,7 +24,7 @@ class QueryBuilderTest extends TestCase
 
     public function testCreateAdapterByFactory()
     {
-        $this->assertInstanceOf('Volcanus\QueryBuilder\Adapter\\Sqlite\\SqliteQueryBuilder',
+        $this->assertInstanceOf(SqliteQueryBuilder::class,
             QueryBuilder::factory(
                 new PdoDriver(
                     new \PDO('sqlite::memory:')
@@ -33,7 +35,7 @@ class QueryBuilderTest extends TestCase
 
     public function testCreateFacade()
     {
-        $this->assertInstanceOf('Volcanus\QueryBuilder\Facade',
+        $this->assertInstanceOf(Facade::class,
             QueryBuilder::facade(
                 new PdoDriver(
                     new \PDO('sqlite::memory:')
